Sign up updated.

+++ src/main/java/com/workfront/internship/workflow/controller/UserController.java
@@ -89,6 +89,12 @@ public class UserController {
 
         String verificationCode = ServiceUtils.hashString(user.getPassword()).substring(0, 6);
 
+        if (code.isEmpty()){
+            redirectAttributes.addFlashAttribute(PageAttributes.MESSAGE,
+                    "Your sign up was canceled.");
+            return new ModelAndView("redirect:/signup");
+        }
+
         if (!code.equals(verificationCode)) {
             userService.deleteById(user.getId());
             redirectAttributes.addFlashAttribute(PageAttributes.MESSAGE,
